fix: sliced design
