--- frontend/educational_article/educational_article.html
@@ -1,157 +1,0 @@
-<!doctype html>
-<!-- saved from url=(0033)http://hightolow.ru/resistor3.php -->
-<html data-darkreader-mode="dynamic" data-darkreader-scheme="dark">
-  <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    <title>Последовательное и параллельное соединение резисторов</title>
-    <link
-      rel="stylesheet"
-      id="hightolow-style-css"
-      href="./educational_article_files/style.css"
-      type="text/css"
-      media="all"
-    />
-  </head>
-  <body>
-    <div style="background: white; width: 75%; margin: auto; padding: 25px">
-      <h1>Последовательное и параллельное соединение резисторов</h1>
-
-      <h2>Последовательное соединение резисторов</h2>
-      <p>
-        <strong>Последовательное соединение</strong> –
-        <em
-          >это соединение двух или более резисторов в форме цепи, в которой каждый отдельный резистор соединяется с
-          другим отдельным резистором только в одной точке.
-        </em>
-      </p>
-
-      <iframe src="https://physics.efedorov.spb.su/embed" width="400px"></iframe>
-      <h2>Общее сопротивление R<sub>общ</sub></h2>
-      <p>
-        При таком соединении, через все резисторы проходит один и тот же электрический ток. Чем больше элементов на
-        данном участке электрической цепи, тем «труднее» току протекать через него. Следовательно, при последовательном
-        соединении резисторов их общее сопротивление увеличивается, и оно равно сумме всех сопротивлений.
-      </p>
-      <img class="formulaImg" id="RtotalSeq" src="./educational_article_files/RtotalSeq.gif" />
-      <h2>Напряжение при последовательном соединении</h2>
-      <p>Напряжение при последовательном соединении распределяется на каждый резистор согласно закону Ома:</p>
-      <img class="formulaImg" id="VRnSeq" src="./educational_article_files/VRnSeq.gif" /> <br />
-      <iframe src="https://physics.efedorov.spb.su/embed" width="400px"></iframe>
-      <p>Т.е чем большее сопротивление резистора, тем большее напряжение на него падает.</p>
-      <h2>Параллельное соединение резисторов</h2>
-      <p>
-        <strong>Параллельное соединение</strong> –
-        <em
-          >это соединение, при котором резисторы соединяются между собой обоими контактами. В результате к одной точке
-          (электрическому узлу) может быть присоединено несколько резисторов.</em
-        >
-      </p>
-
-      <iframe src="https://physics.efedorov.spb.su/embed" width="400px" style="align-self: center"></iframe>
-      <h2>Общее сопротивление R<sub>общ</sub></h2>
-      <p>
-        При таком соединении, через каждый резистор потечет отдельный ток. Сила данного тока будет обратно
-        пропорциональна сопротивлению резистора. В результате общая проводимость такого участка электрической цепи
-        увеличивается, а общее сопротивление в свою очередь уменьшается.
-      </p>
-      <p>
-        <em
-          >Таким образом, при параллельном подсоединении резисторов с разным сопротивлением, общее сопротивление будет
-          всегда меньше значения самого маленького отдельного резистора.
-        </em>
-      </p>
-      <p>Формула общей проводимости при параллельном соединении резисторов:</p>
-      <img
-        class="formulaImg"
-        id="GtotalPar"
-        src="./educational_article_files/GtotalPar.gif"
-        title="общая проводимость при параллельном соединении резисторов"
-      />
-      <p>Формула эквивалентного общего сопротивления при параллельном соединении резисторов:</p>
-      <img
-        class="formulaImg"
-        id="RtotalPar"
-        src="./educational_article_files/RtotalPar.gif"
-        title="общее сопртивление при параллельном соединении резисторов"
-      />
-      <p>Для двух одинаковых резисторов общее сопротивление будет равно половине одного отдельного резистора:</p>
-      <img
-        class="formulaImg"
-        id="RtotalParEqual2R"
-        src="./educational_article_files/RtotalParEqual2R.gif"
-        title="общее сопртивления для двух одинаковых резисторов при параллельном соединении"
-      />
-      <p>
-        Соответственно, для n одинаковых резисторов общее сопротивление будет равно значению одного резистора,
-        разделенного на n.
-      </p>
-      <img
-        class="formulaImg"
-        id="RtotalParEqualNR"
-        src="./educational_article_files/RtotalParEqualNR.gif"
-        title="общее сопртивление одинаковых резисторов при параллельном соединении"
-      />
-      <h2>Напряжение при параллельном соединении</h2>
-      <p>
-        Напряжение между точками A и B является как общим напряжением для всего участка цепи, так и напряжением,
-        падающим на каждый резистор в отдельности. Поэтому при параллельном соединении на все резисторы упадет
-        одинаковое напряжение.
-      </p>
-
-      <iframe src="https://physics.efedorov.spb.su/embed" width="400px" style="align-self: center"></iframe>
-      <h2>Электрический ток при параллельном соединении</h2>
-      <p>
-        Через каждый резистор течет ток, сила которого обратно пропорциональна сопротивлению резистора. Для того чтобы
-        узнать какой ток течет через определенный резистор, можно воспользоваться законом Ома:
-      </p>
-      <img
-        class="formulaImg"
-        id="CurrentPar"
-        src="./educational_article_files/CurrentPar.gif"
-        title="Электрический ток при параллельном соединении"
-      />
-      <br />
-      <h2 id="resistanceMix">Смешанное соединение резисторов</h2>
-      <p>
-        Смешанным соединением называют участок цепи, где часть резисторов соединяются между собой последовательно, а
-        часть параллельно. В свою очередь, смешанное соединение бывает последовательного и параллельного типов.
-      </p>
-
-      <iframe src="https://physics.efedorov.spb.su/embed" width="400px" style="align-self: center"></iframe>
-      <h2>Общее сопротивление R<sub>общ</sub></h2>
-      <p>Для того чтобы посчитать общее сопротивление смешанного соединения:</p>
-      <ul class="articleList">
-        <li>Цепь разбивают на участки с только пареллельным или только последовательным соединением.</li>
-        <li>Вычисляют общее сопротивление для каждого отдельного участка.</li>
-        <li>Вычисляют общее сопротивление для всей цепи смешанного соединения.</li>
-      </ul>
-      <p>Так это будет выглядеть для схемы 1:</p>
-
-      <iframe src="https://physics.efedorov.spb.su/embed" width="400px" style="align-self: center"></iframe>
-      <p>
-        Также существует более быстрый способ расчета общего сопротивления для смешанного соединения. Можно, в
-        соответствии схеме, сразу записывать формулу следующим образом:
-      </p>
-      <ul class="articleList">
-        <li>Если резисторы соединяются последоватеьно — складывать.</li>
-        <li>Если резисторы соединяются параллельно — использовать условное обозначение «||».</li>
-        <li>Подставлять формулу для параллельного соединения где стоит символ «||».</li>
-      </ul>
-      <p>Так это будет выглядеть для схемы 1:</p>
-      <img
-        class="formulaImg"
-        id="RtotalMix"
-        src="./educational_article_files/RtotalMix.gif"
-        title="Общее сопротивление при смешанном соединении. Сокращенная запись"
-      />
-      <p>После подстановки формулы параллельного соединения вместо «||»:</p>
-      <img
-        class="formulaImg"
-        id="RtotalMixLong"
-        src="./educational_article_files/RtotalMixLong.gif"
-        title="Общее сопротивление при смешанном соединении"
-      />
-      <!--Horizontal Separating Bar-->
-    </div>
-  </body>
-</html>
