+++ frontend/src/widgets/canvas/index.ts
@@ -1,2 +1,3 @@
 export { CanvasPanel } from "./ui/canvas-panel.tsx";
 export { Canvas } from "./ui/canvas.tsx";
+export { EducationArticleWidget } from "./ui/article.tsx";
